mobile indexer template & config

+++ poc_iot_injector/src/receipt_txn.rs
@@ -1,5 +1,5 @@
-use crate::{keypair::Keypair, Error, Result};
-use helium_crypto::Sign;
+use crate::{Error, Result};
+use helium_crypto::{Keypair, Sign};
 use helium_proto::{
     services::poc_lora::{
         LoraBeaconReportReqV1, LoraValidBeaconReportV1, LoraValidPocV1, LoraValidWitnessReportV1,
@@ -119,7 +119,7 @@ fn construct_poc_witnesses(
             let poc_witness = BlockchainPocWitnessV1 {
                 gateway: witness_pub_key,
                 timestamp: witness_timestamp,
-                signal: witness_signal as i32,
+                signal: witness_signal,
                 packet_hash: witness_packet,
                 signature: witness_signature,
                 snr: witness_snr as f32,
